Error reporting in -nogui
A lot of warnings were being issued and trying to use the default window
which caused a null pointer when running -nogui. This is fixed by a) not
trying to access frames which are null and b) changing the PAMWarning
system to write messages to terminal rather than open a dialog.