Merge from DG (#77)
* Work on bug fixes in CTD classifier

* couple of final neatenings up of CTD

* DWV Import

Fixed DWV importer so that a binary file is still created if a DWV file
doesn't exist. this is totally normal in quiet conditions, so the
correct thing to do is to generate a binary file with no data so that
effort is correctly recorded.

* Error reporting in -nogui

A lot of warnings were being issued and trying to use the default window
which caused a null pointer when running -nogui. This is fixed by a) not
trying to access frames which are null and b) changing the PAMWarning
system to write messages to terminal rather than open a dialog.