Fixes issues #111 and fixes #112 (DIFAR module crashes and ability to use Deep Learning Detections in DIFAR module) (#110)
* Bugfix for OverlayMarks

Check for a null pointer exception in OverlayMarks that was causing a crash on startup.

* Bugfix for null pointer in symbol manager

Fix a bug that I found where the DIFAR module was crashing the symbol manager. Seems that this was due to this subclass of clip generator having a null value for it's uniqueName. I've fixed by checking for null values and assigning a generic symbol when null.

* DeepLearning detections canGenerateClips=true

Set flag in DeepLearning detector so that detections are considered 'clips' by Pamguard. This allows them to be processed automatically in the DIFAR Localisation module (and maybe others).

* DIFAR: bugfix frequency limits for auto-detections

Fix a bug in DIFAR module where the frequency limits of automated detections were not being set properly by the DIFAR module.

* DeepLearning - Bugfix to detection duration

Fix bug in deep learning detector where duration (in samples) was being set to number of samples in a hop instead of the number of samples in a segment.