Fixed bug in loading viewer data
Fixed a bug where, when getOfflineData was called it was resetting the cancel flag in OfflineDataLoadInfo. This meant that, ifthe load was passed up a chain of modules, then the cancel flag would be reset and a cancel request would be...cancelled. So viewer load requests remained in the queue and were executed after then cancel was set resulting in strange behaviour in the spectrogram display

Also updated X3 to have a public loadSudMap function.