Bug fixes for corrupt wav files and folder issue with sud files
Fixed wav files freezing when header is corrupt on a large file.

Create a more sophisticated file duplicate checker which makes sure that there are no wav and sud files of the same name, even if they are in different sub folders.

Made sure PAMGuard also checks sud map serial UID to bring up worker dialog if sud map needs re-written as well as if it needs created in the first place.