Fix to make sure file times are used when processing sound files
PAMGuard was using a cumulative sample clock to measure time, even if use contiguous files was not set. the sample clock would reset to the file name time if it varied by > 2 seconds. This meant PAMGuard would go out of sync with file times and PAM Binary file time stamps could be up to 2 seconds out. Changed so that the clock is reset to the file start time at the start of every file. One line of code was added to FileInputSystem.prepareSystem.